+++ src/main/java/hello/hellospring/repository/MemoryMemberRepository.java
@@ -7,8 +7,8 @@ import java.util.*;
 
 @Repository
 public class MemoryMemberRepository implements MemberRepsitory {
-    private static Map<Long, Member> store = new HashMap<>();
-    private static long sequence = 0L;
+    private static Map<Integer, Member> store = new HashMap<>();
+    private static Integer sequence = 0;
 
 
     @Override
@@ -19,7 +19,7 @@ public class MemoryMemberRepository implements MemberRepsitory {
     }
 
     @Override
-    public Optional<Member> findById(Long id) {
+    public Optional<Member> findById(Integer id) {
         return Optional.ofNullable(store.get(id));
     }
